Refactor Tuya Air quality
